Bugfix: string token too long in function sym
Every argument to any hin function is tested to find attributes. But
there's no point on testing strings beyond our longest attributes. So I
build a constant setting the longest one and test against it later.

Besides if you give a string to long to sym, it produces an error.